day 11 - change to long

+++ 2020/IRobotElf.cs
@@ -4,8 +4,8 @@ using System.Linq;
 
 interface IRobotElf 
 {
-    int Part1();
-    int Part2();
+    long Part1();
+    long Part2();
 }
 
 abstract class RobotElf : IRobotElf
@@ -20,7 +20,7 @@ abstract class RobotElf : IRobotElf
 
     protected IEnumerable<int> InputNumbers => File.ReadLines($"input/{Day}.txt").Select(int.Parse);
 
-    public abstract int Part1();
+    public abstract long Part1();
 
-    public abstract int Part2();
+    public abstract long Part2();
 }
